Adjusted implicit casting of par values (bug in BMacLayer.cc where the "= hasPar() ? par() : int" construct was used and the compiler does a implicit int cast instead of double). Adjusted also the .ned settings for integer values which was defined as double values. (see https://sourceforge.net/tracker/index.php?func=detail&aid=3496931&group_id=176959&atid=879320)

+++ src/base/modules/BaseMacLayer.h
@@ -73,7 +73,7 @@ protected:
     /**
      * @brief Length of the MacPkt header
      **/
-    int headerLength;
+    long headerLength;
 
     /**
      * @brief MAC address.
@@ -90,7 +90,7 @@ protected:
      * the packet and therefore needs the length of the
      * phy header.
      */
-    int phyHeaderLength;
+    long phyHeaderLength;
 
 private:
     /** @brief Copy constructor is not allowed.
